Start sqlite based implementation
- Add stubs for full dynamodbiface
- Add basic CreateTable handler

The implementation will create a sqlite table per dynamodb table.
The primary row key is a combination of hash(row_key)+row_key. The
goal here is to get hash ordering semantics when doing a table scan.

If you have a range_key we create a sqlite secondary index on that.

We're not handling local or global secondary indexes yet.

We've got a couple of metadata tables:
- __donutdb_global_state: right now just stores table schema version
- __donutdb_table_metadata: additional per table metadata

The table schema version might be useful if we ever decide to use a
better schema system and want to be able to migrate.