Implement read and write batch operations
If we need to do multiple inserts for WriteAt operations, we now do
them in one or more BatchWriteItem operations.

Likewise, when we're doing sequential scans of a file, we now will
read multiple sectors at one. The API exposed is an iterator so the
implementation is mostly hidden from the callers.