Add back a separate row for file metadata
This currently serves two purposes:

It allows us to add a level of indirection between the filename and
the file data. This makes things like `rm file` safer, as we can now
delete the reference to the file and then we don't need to worry too
much about errors occurring while deleting the data (we still care
about deleting the file data so we don't leak rows).

It also allows us to track the sector size on a per file basis. I'm
not sure yet what the optimal sector size is. I'd like to be able to
experiment with different sizes without making old files inaccessible.
It is not currently possible to change the sector size of an existing
file.