Store data and lock keys in metadata
This makes things a bit more inode like. We could implement rename
or maybe hard-links with this without needing to move data
around (hard-links would require a reference counter as well).

The name is stored in the row key to make it easier to inspect the
dynamodb table. If our tooling becomes a bit more robust we could
remove the name from the key.