Compress sectors using zstd
From testing against production dynamodb tables, attribute size makes
a large difference in query latency. The data that I'm working with is
mostly text and compresses well. Also, at least with young sqlite
database files, they start off fairly sparse.

We store the compression algorithm in the file metadata so if we want
to upgrade the algorithm at some point in the future we can do so
easily.