Make FullPathname always prefix paths with '/'
This mainly makes it easier to work with the loadable module from the
sqlite3 cli and the sqlite URI syntax. The URI syntax is always an
absolute path so will always begin with the path separator.

Without this, you can accidentally create db files in dynamo that you
would be unable to access from the loadable module.

Closes: #7 [via git-merge-pr]