Change how our sector fill logic works.
We now fill all sectors if we are growing a file. Before we just
filled the empty sectors. With the error changes to the sector
iterator, this code was the only place where we were trying to
iterate over sectors that didn't exist.

It was simpler to initialize those sectors than to change the fill
logic. I also am suspicious that this code was the source of some
existing bugs.